Add files via upload
Machine learning is a branch of artificial intelligence (AI) that enables computer systems to learn and improve from experience without being explicitly programmed. It involves the development of algorithms and models that allow computers to automatically learn patterns and make decisions based on data. There are various types of machine learning. The two main types of machine learning are supervised and unsupervised learning.